Converted all values to BigDecimal, removed Map from Events. Here goes nothing.

+++ src/main/java/com/localstrategy/Strategy2.java
@@ -4,6 +4,7 @@ import com.localstrategy.util.types.Candle;
 import com.localstrategy.util.types.Position;
 import com.localstrategy.util.types.SingleTransaction;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -36,8 +37,8 @@ public class Strategy2 {
     public void candleUpdate(Candle lastCandle){
         if(lastCandle.tick() <= -distance){ // New lows for long
             packingLow = true;
-            if(lastCandle.low() < stopPriceLow){
-                stopPriceLow = lastCandle.low();
+            if(lastCandle.low().doubleValue() < stopPriceLow){
+                stopPriceLow = lastCandle.low().doubleValue();
             }
         } else if (packingLow){
 
@@ -45,7 +46,7 @@ public class Strategy2 {
                 handler.closePosition(position);
             }
 
-            handler.activateStopLoss(handler.executeMarketOrder(stopPriceLow));
+            handler.activateStopLoss(handler.executeMarketOrder(BigDecimal.valueOf(stopPriceLow)));
 
             packingLow = false;
             stopPriceLow = Double.MAX_VALUE;
@@ -53,8 +54,8 @@ public class Strategy2 {
 
         if(lastCandle.tick() >= distance){ // New highs for short
             packingHigh = true;
-            if(lastCandle.high() > stopPriceHigh){
-                stopPriceHigh = lastCandle.high();
+            if(lastCandle.high().doubleValue() > stopPriceHigh){
+                stopPriceHigh = lastCandle.high().doubleValue();
             }
         } else if (packingHigh){
 
@@ -62,7 +63,7 @@ public class Strategy2 {
                 handler.closePosition(position);
             }
 
-            handler.activateStopLoss(handler.executeMarketOrder(stopPriceHigh));
+            handler.activateStopLoss(handler.executeMarketOrder(BigDecimal.valueOf(stopPriceHigh)));
 
             packingHigh = false;
             stopPriceHigh = 0;
